finish: settings page

+++ src/pages/Dashboard/components/Video.tsx
@@ -2,38 +2,49 @@ import { AxiosError } from 'axios';
 import toast from 'react-hot-toast';
 import { Tooltip } from '@mui/material';
 import { useDispatch } from 'react-redux';
+import { useAuth0 } from '@auth0/auth0-react';
 import { useNavigate } from 'react-router-dom';
-import { faTrashCan } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faRecycle, faTrashCan } from '@fortawesome/free-solid-svg-icons';
 import React, { memo, useEffect, useMemo, useRef, useState } from 'react';
 
+import { AppDispatch } from '../../../store';
 import constants from '../../../static/constants.json';
 import { fancyTimeFormat, trimStr } from '../../../common/helpers';
-import MemoziedCircularProgress from '../../../components/CircularProgress';
 import { VideoState } from '../../../store/reducers/videos/index.interface';
-import { deleteVideo, getAllVideos, getThumbnail } from '../../../api/videos';
+import { fetchAllVideos } from '../../../store/reducers/videos/index.thunk';
 import MemoziedConfirmationModal from '../../../components/DeleteConfirmationModal';
+import MemoziedCircularProgress from '../../../components/CircularProgressWithLabel';
 import {
     deleteByVideoId,
     updateVideoProcessingState,
 } from '../../../store/reducers/videos';
+import {
+    deleteVideo,
+    getAllVideos,
+    getThumbnail,
+    reIndexVideo,
+} from '../../../api/videos';
 
 interface VideoProps {
     video: VideoState;
 }
 
 const Video: React.FC<VideoProps> = ({ video }) => {
+    const { isAuthenticated } = useAuth0();
     const { id, thumbnailId, processingProgress, state, name } = video;
 
     const shouldRender = useRef(true);
     const [thumbnail, setThumbnail] = useState('');
     const [isDeleteProcessing, setIsDeleteProcessing] = useState(false);
+    const [isReIndexing, setIsReIndexing] = useState(true);
     const [invalidImageURLError, setInvalidImageURLError] = useState(false);
 
-    const dispatch = useDispatch();
+    const dispatch = useDispatch<AppDispatch>();
     const navigate = useNavigate();
 
     const isProcessedVideo = useMemo(
-        () => state?.toLowerCase() === constants.VIDEO_PROCESSED,
+        () => state?.toLowerCase() === constants.VIDEOS.VIDEO_PROCESSED,
         [state]
     );
 
@@ -51,15 +62,41 @@ const Video: React.FC<VideoProps> = ({ video }) => {
     ): Promise<void> => {
         event.stopPropagation();
         try {
-            if (isProcessedVideo) {
+            if (isProcessedVideo && isAuthenticated) {
                 setIsDeleteProcessing(true);
                 await deleteVideo(id);
                 dispatch(deleteByVideoId({ videoId: id }));
-                setIsDeleteProcessing(false);
                 toast.success(constants.SUCCESS_MESSAGE.VIDEO_DELETED);
+            } else if (!isAuthenticated) {
+                toast.error(constants.ERROR_MESSAGE.GUEST_RESTRICTION);
             }
+            setIsDeleteProcessing(false);
         } catch (error: unknown) {
             setIsDeleteProcessing(false);
+            if (error instanceof Error) toast.error(error?.message);
+            else if (error instanceof AxiosError)
+                toast.error(error?.response?.data?.message);
+        }
+    };
+
+    const handleReIndexVideo = async (
+        event: React.MouseEvent<HTMLButtonElement | HTMLOrSVGElement>
+    ): Promise<void> => {
+        event?.stopPropagation();
+        try {
+            if (isProcessedVideo && isAuthenticated) {
+                setIsReIndexing(true);
+                const response = await reIndexVideo(video.id);
+                if (response === 'ok') {
+                    toast.success(constants.SUCCESS_MESSAGE.REINDEX_INITIATED);
+                    dispatch(fetchAllVideos());
+                }
+                setIsReIndexing(false);
+            } else if (!isAuthenticated) {
+                toast.error(constants.ERROR_MESSAGE.GUEST_RESTRICTION);
+            }
+        } catch (error: unknown) {
+            setIsReIndexing(false);
             if (error instanceof Error) toast.error(error?.message);
             else if (error instanceof AxiosError)
                 toast.error(error?.response?.data?.message);
@@ -118,7 +155,7 @@ const Video: React.FC<VideoProps> = ({ video }) => {
                     else if (error instanceof AxiosError)
                         toast.error(error?.response?.data?.message);
                 }
-            }, constants.VIDEO_PROCESSING_API_CALL);
+            }, constants.VIDEOS.VIDEO_PROCESSING_API_CALL);
 
         return () => clearInterval(intervalId);
     }, [dispatch, thumbnail, id, isProcessedVideo]);
@@ -155,13 +192,26 @@ const Video: React.FC<VideoProps> = ({ video }) => {
                     <p className="bg-darkgrey absolute text-xs lg:text-sm bottom-4 lg:bottom-2 font-cherry right-2 bg-opacity-75 rounded px-2 py-1 cursor-default">
                         {fancyTimeFormat(video?.durationInSeconds)}
                     </p>
+                    <button
+                        type="button"
+                        onClick={handleReIndexVideo}
+                        disabled={isReIndexing}
+                        className="bg-darkgrey absolute text-xs bg-opacity-75 rounded px-2 py-1 top-4 lg:top-2 right-12 hover:scale-95 cursor-pointer"
+                    >
+                        <Tooltip title="Re-index">
+                            <FontAwesomeIcon
+                                icon={faRecycle}
+                                onClick={handleReIndexVideo}
+                            />
+                        </Tooltip>
+                    </button>
                 </>
             ) : (
                 <div className="absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 flex items-center justify-center flex-col">
                     <MemoziedCircularProgress
                         value={parseInt(processingProgress)}
                     />
-                    {constants.VIDEO_INDEXING_MESSAGE}
+                    {constants.VIDEOS.VIDEO_INDEXING_MESSAGE}
                 </div>
             )}
         </div>
